fix deleted all attachments on query

+++ src/controller/transaction/transaction.service-v2.ts
@@ -1017,6 +1017,7 @@ export class TransactionService {
   }
   public async deleteAttachmentByTransaction(id: string, tx: Prisma.TransactionClient) {
     try {
+      if (!id) throw new Error("No ID provided ! ");
       await tx.attachment.deleteMany({
         where: {
           transactionId: id,
